added keypad with password
added a keypad that functions. If there is smoke detected, the light goes flashing like a disco. When the '*' is pressed the code waits for a number input of 4. When correct the alarm stops. When incorrect it messages the user that the password is wrong and the alarm keeps going